Test data, docker, integration tests

+++ src/pav3/align/score.py
@@ -123,7 +123,7 @@ class ScoreModel(ABC):
 
         :returns: Sum of alignment scores across all operations.
         """
-        return np.sum(np.vectorize(self.score_op)(op_arr[:, 0], op_arr[:, 1]))
+        return float(np.sum(np.vectorize(self.score_op)(op_arr[:, 0], op_arr[:, 1])))
 
     def score_align_table(self, df: pl.DataFrame) -> pl.Series:
         """Score all alignment records in a table.
@@ -138,7 +138,7 @@ class ScoreModel(ABC):
             .select(pl.col('align_ops'))
             .to_series()
             .map_elements(op.row_to_arr, return_dtype=pl.Object)
-            .map_elements(self.score_op_arr, return_dtype=pl.Float32)
+            .map_elements(self.score_op_arr, return_dtype=pl.Float64)
         )
 
         # A pure polars implementation is slower:
